Modified drill1 to use double

+++ Cpp/programming-book/chapter4/drill1.cpp
@@ -2,13 +2,13 @@
 #include<vector>
 
 using namespace std;
-void print_values(int,int);
+void print_values(double,double);
 
 int main()
 {
-  vector<int> nums;
-  int num1=0;
-  int num2=0;
+  vector<double> nums;
+  double num1=0;
+  double num2=0;
   while (cin>>num1>>num2)
   {
    cout << num1 << '\t' << num2 << '\n';
@@ -27,7 +27,7 @@ int main()
 
 
 
-void print_values(int num1, int num2)
+void print_values(double num1, double num2)
 {
      cout << "the smaller value is: " << num1 << '\n';
      cout << "the larger value is: " << num2 << '\n';
